LayoutTests/plugins/npruntime/remove-property.html


<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<script>
function runTest() {
    plugin = document.getElementById('plugin');

    obj = { property : 'value' }
    shouldBe("'property' in obj", "true");
    plugin.testRemoveProperty(obj, 'property');
    shouldBe("'property' in obj", "false");

    array = [ 1, 2, 3, 4, 5];
    shouldBe("array[1]", "2")
    plugin.testRemoveProperty(array, 1);
    shouldBe("array[1]", "undefined")
}
</script>
<body onLoad="runTest()">
<embed id="plugin" type="application/x-webkit-test-netscape" test="npruntime-remove-property"></embed>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test NPN_RemoveProperty");

</script>
