LayoutTests/plugins/npruntime/remove-property-expected.txt


Test NPN_RemoveProperty

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'property' in obj is true
PASS 'property' in obj is false
PASS array[1] is 2
PASS array[1] is undefined
